[WARP-SPEED] Fix class is exposed outside of visibality scope warnings
(cherry picked from commit acc278762dd81a2921e1351b04048dd6617a6c92)

+++ plugin/starburst-warp-speed/starburst-warp-speed-core/src/main/java/io/trino/plugin/warp/storage/read/ChunksQueue.java
@@ -43,7 +43,7 @@ class ChunksQueue
             if (matchBitmapResetPoints[i] < 0) {
                 throw new TrinoException(WARP_UNRECOVERABLE_MATCH_FAILED, "match got to error state at chunk " + matchedChunksIndexes[i]);
             }
-            chunksToCollect.add(new MatchChunkResult((int) matchedChunksIndexes[i], matchBitmapResetPoints[i]));
+            chunksToCollect.add(new MatchChunkResult(matchedChunksIndexes[i], matchBitmapResetPoints[i]));
         }
         this.totalNumChunks = totalNumChunks;
     }
